requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.2
django-crispy-forms==1.9.2
djangorestframework==3.12.1
google-api-core==1.22.4
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.2.0
google-cloud-bigquery-storage==2.0.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
libcst==0.3.13
mypy-extensions==0.4.3
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
pandas-gbq==0.14.0
proto-plus==1.11.0
protobuf==3.13.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
